% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_package.R
\name{abc}
\alias{abc}
\title{Example dataset}
\format{
A dataframe with 5 variables:
\describe{
  \item{AGE}{Age}
  \item{TRT}{Treatment assignment. 0: UC + SBT, 1: SAT + SBT}
  \item{SURV}{Survival days}
  \item{Y2}{Cognitive score at 12 months}
  \item{Y1}{Cognitive score at 3 months}
}
}
\description{
The Awakening and Breathing Controlled (ABC) trial randomized critically ill
patients receiving mechanical ventilation 1:1 within each study site to
management with a paired sedation plus ventilator weaning protocol involving
daily interruption of sedative through spontaneous awakening trials (SATs)
and spontaneous breathing trials (SBTs) or sedation per usual care
(UC) and SBTs.

The example dataset is from a single site substudy in ABC. The researchers
assessed differences in cognitive, psychological and functional outcomes at 3
and 12 months after randomization.
}
\references{
T. D. Girard, J. P. Kress, B. D. Fuchs, J. W. W. Thomason, W. D. Schweickert,
B. T. Pun, D. B. Taichman, J. G. Dunn, A. S. Pohlman, P. A. Kinniry, J. C.
Jackson, A. E. Canonico, R. W. Light, A. K. Shintani, J. L. Thompson, S. M.
Gordon, J. B. Hall, R. S. Dittus, G. R. Bernard, and E. W. Ely. Efficacy and
safety of a paired sedation and ventilator weaning protocol for mechanically
ventilated patients in intensive care (awakening and breathing controlled
trial): a randomised controlled trial. Lancet, 371:126-134, 2008.
}
