% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_visualize.R
\name{chkPara}
\alias{chkPara}
\title{Check parameter specification}
\usage{
chkPara(lst.var, data.all, html = FALSE)
}
\arguments{
\item{lst.var}{see \code{\link{idem-parameters}}}

\item{data.all}{Original dataset}

\item{html}{logic indicator for the format of the error messages}
}
\value{
Text messages if \code{html=TRUE} .
}
\description{
Check if the \code{idem-parameters} are correctly specified and consistent
with the data
}
\examples{

err.lst.var <- list(trt="TRT", outcome=c("Y1","Y2"),
                y0=NULL, endfml="Y3", bounds=c(10,20),
                duration=365);

chkPara(err.lst.var, abc);

}

