% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem-package.R
\docType{package}
\name{idem-package}
\alias{idem}
\alias{idem-package}
\title{idem: XLSForm Comparison and Validation}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Reads and compares XLSForm survey files, validating that a target form is consistent with a development form.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://impact-initiatives.github.io/idem/}
}

}
\author{
\strong{Maintainer}: Iyed GHEDAMSI \email{iyed.ghedamsi@impact-initiatives.org}

Authors:
\itemize{
  \item Iyed GHEDAMSI \email{iyed.ghedamsi@impact-initiatives.org}
}

}
\keyword{internal}
