% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passing_lists.R
\name{idem_passing_lists}
\alias{idem_passing_lists}
\title{Default list names skipped by \code{\link[=validate_choices]{validate_choices()}}}
\usage{
idem_passing_lists
}
\description{
A character vector of XLSForm list names whose choice options are expected
to differ between forms (admin boundaries, cluster IDs, enumerator IDs).
Used as the default for the \code{passing_lists} argument of
\code{\link[=validate_choices]{validate_choices()}} and \code{\link[=validate_xlsform]{validate_xlsform()}}.
}
\examples{
idem_passing_lists
}
