% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_imputation.R
\name{imChkPars}
\alias{imChkPars}
\title{Check parameter specification}
\usage{
imChkPars(data.all, lst.var, html = FALSE)
}
\arguments{
\item{data.all}{Original dataset}

\item{lst.var}{see \code{\link{idem-parameters}}}

\item{html}{logic indicator for the format of the error messages}
}
\value{
\code{NULL} if the specification is correct.

  Text messages if \code{html=TRUE}.
}
\description{
Check if the \code{idem-parameters} are correctly specified and consistent
with the data
}
\examples{

err.lst.var <- list(trt="TRT", outcome=c("Y1","Y2"),
                y0=NULL, endfml="Y3", bounds=c(10,20),
                duration=365);

imChkPars(abc, err.lst.var);

}

