% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_analysis.R
\name{imInfer2}
\alias{imInfer2}
\title{Treatment effect estimation and hypothesis testing}
\usage{
imInfer2(imp.rst, test.rst, effect.quantiles = 0.5, quant.ci = c(0.025,
  0.975))
}
\description{
Estimation and hypothesis testing based on imputation and bootstrap results
}
