% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_visualize.R
\name{imMisTable}
\alias{imMisTable}
\title{Generate table of missingness pattern frequencies}
\usage{
imMisTable(data.all, lst.var)
}
\arguments{
\item{data.all}{Original dataset}

\item{lst.var}{see \code{\link{idem-parameters}}}
}
\value{
A matrix with frequencies of each missing pattern
}
\description{
Generate table of missingness pattern frequencies
}
\examples{

lst.var <- list(trt="TRT", surv="SURV", outcome=c("Y1","Y2"),
                y0=NULL, trt.label = c("UC+SBT", "SAT+SBT"),
                duration=365);

imMisTable(abc, lst.var);

}

