% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_shiny.R
\name{imShiny}
\alias{imShiny}
\title{Run Web-Based \code{idem} application}
\usage{
imShiny()
}
\description{
Call Shiny to run \code{idem} as a web-based application. A web browser will
be brought up.
}
\examples{
\dontrun{
run.idem()}


}

