% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_visualize.R
\name{plotCompleters}
\alias{plotCompleters}
\title{Plot data of completer}
\usage{
plotCompleters(data.all, lst.var, fname = NULL, ...)
}
\arguments{
\item{data.all}{Original dataset}

\item{lst.var}{see \code{\link{idem-parameters}}}

\item{fname}{File name of the result pdf file. If \code{fname} is null,
result pdf file will not be generated}

\item{...}{Options for \code{pdf} function}
}
\description{
Spaghetti plot for subjects alive at the end of the study without missing data
}
\examples{

lst.var <- list(trt="TRT", surv="SURV", outcome=c("Y1","Y2"),
                y0=NULL, trt.label = c("UC+SBT", "SAT+SBT"),
                duration=365);

plotCompleters(abc, lst.var);

}

