% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_visualize.R
\name{plotMisPattern}
\alias{plotMisPattern}
\title{Plot missing patterns}
\usage{
plotMisPattern(data.all, lst.var, cols = c("blue", "gray"), fname = NULL,
  ...)
}
\arguments{
\item{data.all}{Original dataset}

\item{lst.var}{see \code{\link{idem-parameters}}}

\item{cols}{Color of observed and missing values}

\item{fname}{File name of the result pdf file. If \code{fname} is null,
result pdf file will not be generated}

\item{...}{Options for \code{pdf} function}
}
\description{
Plot the missing patterns of the observed data
}
\examples{

lst.var <- list(trt="TRT", outcome=c("Y1","Y2"),
                trt.label = c("UC+SBT", "SAT+SBT"));

plotMisPattern(abc, lst.var);

}

