% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_visualize.R
\name{plotSurv}
\alias{plotSurv}
\title{Plot survival curves}
\usage{
plotSurv(data.all, lst.var, cols = c("black", "blue"), fname = NULL, ...)
}
\arguments{
\item{data.all}{Original dataset}

\item{lst.var}{see \code{\link{idem-parameters}}}

\item{cols}{Curve colors of the treatment and control arm}

\item{fname}{File name of the result pdf file. If \code{fname} is null,
result pdf file will not be generated}

\item{...}{Options for \code{pdf} function}
}
\description{
Plot Kaplan-Meier survival curves
}
\examples{

lst.var <- list(trt="TRT", surv="SURV", outcome=c("Y1","Y2"),
                y0=NULL, trt.label = c("UC+SBT", "SAT+SBT"),
                duration=365);
plotSurv(abc, lst.var);

}

