% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_data.R
\name{print.IDEMDATA}
\alias{print.IDEMDATA}
\title{Print IDEMDATA object}
\usage{
\method{print}{IDEMDATA}(x, ...)
}
\arguments{
\item{x}{A class \code{IDEMDATA} object generated by \code{\link{imData}}}

\item{...}{Additional arguments}
}
\description{
Print the specification details of class \code{IDEMDATA} objects generated by
\code{\link{imData}}
}
\seealso{
\code{\link{imData}}
}
