% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_data.R
\name{print.IDEMERROR}
\alias{print.IDEMERROR}
\title{Print error messages}
\usage{
\method{print}{IDEMERROR}(x, html = FALSE, ...)
}
\arguments{
\item{x}{A class \code{IDEMERROR} object returned by
\code{\link{imData}} when there are misspecifications}

\item{html}{Logical indicator for the format of the error messages. When \code{TRUE},
the error messages are formatted in HTML format}

\item{...}{Additional arguments}
}
\description{
Print error messages in the parameter specifications generated by
\code{\link{imData}}
}
\examples{
\dontrun{
rst.data <- imData(abc, trt="TRT", outcome=c("Y1","Y2"), y0=NULL,
                   endfml="Y3", bounds=c(10,20), duration=365,
                   err.terminate=FALSE);
print(rst.data);}

}
\seealso{
\code{\link{imData}}
}
