% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_imputation.R
\name{print.IDEMFIT}
\alias{print.IDEMFIT}
\title{Print model fitting results}
\usage{
\method{print}{IDEMFIT}(x, ...)
}
\arguments{
\item{x}{A class \code{IDEMFIT} object generated by \code{\link{imFitModel}}}

\item{...}{Additional arguments}
}
\description{
Print method of the class \code{IDEMFIT} generated by
\code{\link{imFitModel}}
}
\details{
Print the results from \code{lm} for all the models
}
\seealso{
\code{\link{imFitModel}}
}
