% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_imputation.R
\name{print.IDEMIMP}
\alias{print.IDEMIMP}
\title{Print imputation results}
\usage{
\method{print}{IDEMIMP}(x, ...)
}
\arguments{
\item{x}{A class \code{IDEMIMP} object returned from \code{\link{imImpAll}}}

\item{...}{Extra arguments}
}
\description{
Print method for class \code{IDEMIMP} objects generated
by \code{\link{imImpAll}}
}
\seealso{
\code{\link{imImpAll}}
}
