% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_imputation.R
\name{print.IDEMSINGLE}
\alias{print.IDEMSINGLE}
\title{Print MCMC mixing checking result}
\usage{
\method{print}{IDEMSINGLE}(x, ...)
}
\arguments{
\item{x}{A class \code{IDEMSINGLE} object returned from \code{\link{imImpSingle}}}

\item{...}{Additional arguments}
}
\description{
Print method for class \code{IDEMSINGLE} objects generated
by \code{\link{imImpSingle}}
}
\seealso{
\code{\link{imImpSingle}}
}
