% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_sace.R
\name{print.summary.IDEMINFER}
\alias{print.summary.IDEMINFER}
\title{Print survivors only or SACE analysis results}
\usage{
\method{print}{summary.IDEMINFER}(x, delta0 = NULL, delta1 = NULL, ...)
}
\arguments{
\item{x}{A class \code{IDEMINFER} list generated by \code{\link{imInfer}}}

\item{delta0}{Selected treatment arm 0 sensitivity parameters}

\item{delta1}{Selected treatment arm 1 sensitivity parameters}

\item{...}{Extra arguments}
}
\description{
Print survivors only or SACE analysis results
}
