% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idem_shiny.R
\name{imShiny}
\alias{imShiny}
\title{Run Web-Based \code{idem} application}
\usage{
imShiny()
}
\description{
Call Shiny to run \code{idem} as a web-based application.
}
\details{
A web browser will be brought up for users to access the GUI of \code{idem}.
}
\examples{
\dontrun{
run.idem()}


}
