% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_labels.R
\docType{data}
\name{translatable_fields}
\alias{translatable_fields}
\title{Translatable XLSForm fields}
\format{
An object of class \code{character} of length 8.
}
\usage{
translatable_fields
}
\description{
A character vector of column name prefixes that support per-language
translations in XLSForm. Each field can be suffixed with a language tag
(e.g. \code{label::English (en)}, \code{hint::French (fr)}) following the
\href{https://xlsform.org/en/#multiple-language-support}{XLSForm multiple language convention}.
}
\details{
The parenthesised code at the end of the suffix (e.g. \code{en}, \code{fr}) should be
a valid IANA language subtag. The full registry is available at
\url{https://www.iana.org/assignments/language-subtag-registry}.
}
\examples{
translatable_fields

}
\keyword{datasets}
